added bogus tests for listener & dialer wrappers
